requirements.txt
flask
mysql
python-dotenv
mysql.connector
flask_bcrypt
